requirements.txt
Django==3.2.8
djangorestframework==3.12.4

djangorestframework-simplejwt==4.8.0
pyjwt==2.1.0
drf-yasg==1.20.0 #swagger
django-filter==2.4.0

redis==3.5.3
django-redis==4.12.1
requests
python-dotenv==0.17.0
django-celery-beat==2.2.1
django-cors-headers==3.7.0
loguru==0.6.0


pymupdf==1.20.2
pytest==6.2.3
pytest-cov==2.11.1
pytest-django==4.2.0
pytest-mock==3.5.1
model-bakery==1.2.1
